+++ resources/publishable/resources/views/pages/home.blade.php
@@ -1,7 +1,4 @@
 <x-gt-app-layout layout="{{ config('gotime.template') }}" hasContainer class="py-5-3-2-2">
 
-    <h1>{{ $pageTitle ?? null }}</h1>
-
-
 
 </x-gt-app-layout>