+++ resources/views/components/layouts/partials/footer.blade.php
@@ -1,7 +1,5 @@
-<footer class="nk-footer">
-
+<footer class="nk-footer tac">
     <div class="copyright">
         <small>{{ config('gotime.copyright') }}&trade; Copyright © <?php echo date('Y'); ?> &nbsp;&nbsp; | &nbsp;&nbsp; Design and Hosting by <a href="https://naykel.com.au" target="_blank">NAYKEL</a></small>
     </div>
-
 </footer>