+++ resources/views/components/v2/input/partials/error.blade.php
@@ -1,0 +1,7 @@
+@props(['for'])
+
+@error($for)
+    <p class="mt-025 txt-xs txt-red-600" role="alert">
+        {{ $message }}
+    </p>
+@enderror